DataForecast.TimeSeries

Description

TimeSeries

data TimeSeries parts where

Time series date representation with the resolution breakdown encoded in the type.

parts has kind '[TsPeriod]' which represents the hierarchical breakdown of your data. You are responsible for providing the correct type hierarchy for your data initially, but once you have done that the library ensures that all analysis performed is valid.

In general you should not use the TimeSeries constructor directly, instead use the raw and fromParts construction functions defined below.

Constructors

TimeSeries :: SPeriod p -> SummaryData -> Subparts subparts -> TimeSeries (p ': subparts)

getSD :: TimeSeries parts -> SummaryData

Get the SummaryData out of a TimeSeries.

setSD :: SummaryData -> TimeSeries parts -> TimeSeries parts

Set the SummaryData field of a TimeSeries. We can't use record syntax because https://ghc.haskell.org/trac/ghc/ticket/2595 is not implemented.

getSub :: TimeSeries (p ': subparts) -> Subparts subparts

Get the Subparts out of a TimeSeries.

setSub :: Subparts subparts -> TimeSeries (p ': subparts) -> TimeSeries (p ': subparts)

Set the Subparts of the TimeSeries.

Construction.

class BuildTS p where

Typeclass for constructing TimeSeries with default data.

This is used by several of the TimeSeries construction functions such as raw. It provides a convenient way for constructing time series of different types with the same functions and allowing the concrete type to be pinned down by a type signature.

Minimal complete definition

build

Methods

build :: SummaryData -> Subparts rest -> TimeSeries (p ': rest)

raw :: BuildTS p => Double -> TimeSeries '[p]

Construct a leaf TimeSeries with the given total.

Example:

raw 42 :: TimeSeries '[ 'Year ]

fromParts :: BuildTS part => [TimeSeries subparts] -> TimeSeries (part ': subparts)

Construct a TimeSeries given the constituent TimeSeries parts.

Example:

yearByQuarter :: TimeSeries '[ 'Year, 'Quarter ]
yearByQuarter = fromParts [ raw 10, raw 20, raw 30, raw 15 ]

TsPeriod

data TsPeriod

Use promoted constructors to represent the different timeseries that we have. When using these as kinds you must prefix them with a single quote (').

Example:

raw 4 :: TimeSeries '[ 'Day ]

Constructors

Decade

Year

Quarter

Month

Day

data SPeriod x where

The singleton for the promoted TsPeriod type.

See https://www.schoolofhaskell.com/user/konn/prove-your-haskell-for-great-safety/dependent-types-in-haskell#singleton-patterns for details.

Constructors

SDecade :: SPeriod Decade

SYear :: SPeriod Year

SQuarter :: SPeriod Quarter

SMonth :: SPeriod Month

SDay :: SPeriod Day

SummaryData

data SummaryData

The summarization of the entire time series rooted at the current 'TimeSeries.

Constructors

SummaryData

Fields

sdtotal :: Maybe Double

sdmean :: Maybe Double

defaultSummary :: SummaryData

The default SummaryData with Nothing for the sdtotal and sdmean.

summaryWithTotal :: Double -> SummaryData

Construct a SummaryData with the sdtotal set.

setSdTotal :: Double -> SummaryData -> SummaryData

Set the total for the SummaryData.

setSdMean :: Double -> SummaryData -> SummaryData

Set the mean for the SummaryData.

Subparts

data Subparts parts

The constituent components of a TimeSeries. If the TimeSeries is just a leaf with raw data then subs will be Nothing.

Constructors

Subparts

Fields

subs :: Maybe [TimeSeries parts]