DataForecast.TimeSeries.Aggregate

Description

Documentation

computeTotal :: TimeSeries parts -> TimeSeries parts

Compute the total for each SummaryData in the TimeSeries.

This propogates totals bottom up from the raw leaf data to the root and updating all SummaryData nodes along the way.