DataForecast.TimeSeries.Analysis

Documentation

mean :: TimeSeries (p ': rest) -> Maybe Double

Return the mean of the timeseries.

It is possible for mean to be undefined and return Nothing if given an empty TimeSeries. If that's the case the caller can determine what a reasonable fallback is.

std :: TimeSeries (p ': rest) -> Maybe Double

Compute the standard deviation for the time series across the top dimension.

Note: